wgpu-core: Move registry code to new `registry` module.
