+++ wgpu-core/src/hub.rs
@@ -143,14 +143,15 @@ flagged as errors as well.
 [`Global`]: crate::global::Global
 [`Global::new`]: crate::global::Global::new
 [`gfx_select`]: crate::gfx_select
-[`Input`]: IdentityHandler::Input
-[`process`]: IdentityHandler::process
+[`IdentityHandler`]: crate::identity::IdentityHandler
+[`Input`]: crate::identity::IdentityHandler::Input
+[`process`]: crate::identity::IdentityHandler::process
 [`Id<R>`]: crate::id::Id
-[wrapped in a mutex]: trait.IdentityHandler.html#impl-IdentityHandler%3CI%3E-for-Mutex%3CIdentityManager%3E
+[wrapped in a mutex]: ../identity/trait.IdentityHandler.html#impl-IdentityHandler%3CI%3E-for-Mutex%3CIdentityManager%3E
 [WebGPU]: https://www.w3.org/TR/webgpu/
 [`IdentityManager`]: crate::identity::IdentityManager
 [`Input<G, I>`]: crate::identity::Input
-
+[`IdentityHandlerFactory<I>`]: crate::identity::IdentityHandlerFactory
 */
 
 use crate::{
@@ -159,14 +160,14 @@ use crate::{
     device::Device,
     hal_api::HalApi,
     id,
-    identity::{GlobalIdentityHandlerFactory, IdentityHandler, IdentityHandlerFactory},
+    identity::GlobalIdentityHandlerFactory,
     instance::{Adapter, HalSurface, Instance, Surface},
     pipeline::{ComputePipeline, RenderPipeline, ShaderModule},
+    registry::Registry,
     resource::{Buffer, QuerySet, Sampler, StagingBuffer, Texture, TextureClearMode, TextureView},
     Epoch, Index,
 };
 
-use parking_lot::{RwLock, RwLockReadGuard, RwLockWriteGuard};
 use wgt::Backend;
 
 #[cfg(debug_assertions)]
@@ -215,8 +216,8 @@ pub(crate) struct InvalidId;
 #[derive(Debug)]
 pub struct Storage<T, I: id::TypedId> {
     pub(crate) map: Vec<Element<T>>,
-    kind: &'static str,
-    _phantom: PhantomData<I>,
+    pub(crate) kind: &'static str,
+    pub(crate) _phantom: PhantomData<I>,
 }
 
 impl<T, I: id::TypedId> ops::Index<id::Valid<I>> for Storage<T, I> {
@@ -547,7 +548,7 @@ impl<'a, T> Token<'a, T> {
     /// Return a new token for a locked field.
     ///
     /// This should only be used by `Registry` locking methods.
-    fn new() -> Self {
+    pub(crate) fn new() -> Self {
         #[cfg(debug_assertions)]
         ACTIVE_TOKEN.with(|active| {
             let old = active.get();
@@ -591,188 +592,6 @@ pub trait Resource {
         return &self.life_guard().label;
         #[cfg(not(debug_assertions))]
         return "";
-    }
-}
-
-#[derive(Debug)]
-pub struct Registry<T: Resource, I: id::TypedId, F: IdentityHandlerFactory<I>> {
-    identity: F::Filter,
-    pub(crate) data: RwLock<Storage<T, I>>,
-    backend: Backend,
-}
-
-impl<T: Resource, I: id::TypedId, F: IdentityHandlerFactory<I>> Registry<T, I, F> {
-    fn new(backend: Backend, factory: &F) -> Self {
-        Self {
-            identity: factory.spawn(),
-            data: RwLock::new(Storage {
-                map: Vec::new(),
-                kind: T::TYPE,
-                _phantom: PhantomData,
-            }),
-            backend,
-        }
-    }
-
-    pub(crate) fn without_backend(factory: &F, kind: &'static str) -> Self {
-        Self {
-            identity: factory.spawn(),
-            data: RwLock::new(Storage {
-                map: Vec::new(),
-                kind,
-                _phantom: PhantomData,
-            }),
-            backend: Backend::Empty,
-        }
-    }
-}
-
-#[must_use]
-pub(crate) struct FutureId<'a, I: id::TypedId, T> {
-    id: I,
-    data: &'a RwLock<Storage<T, I>>,
-}
-
-impl<I: id::TypedId + Copy, T> FutureId<'_, I, T> {
-    #[cfg(feature = "trace")]
-    pub fn id(&self) -> I {
-        self.id
-    }
-
-    pub fn into_id(self) -> I {
-        self.id
-    }
-
-    pub fn assign<'a, A: Access<T>>(self, value: T, _: &'a mut Token<A>) -> id::Valid<I> {
-        self.data.write().insert(self.id, value);
-        id::Valid(self.id)
-    }
-
-    pub fn assign_error<'a, A: Access<T>>(self, label: &str, _: &'a mut Token<A>) -> I {
-        self.data.write().insert_error(self.id, label);
-        self.id
-    }
-}
-
-impl<T: Resource, I: id::TypedId + Copy, F: IdentityHandlerFactory<I>> Registry<T, I, F> {
-    pub(crate) fn prepare(
-        &self,
-        id_in: <F::Filter as IdentityHandler<I>>::Input,
-    ) -> FutureId<I, T> {
-        FutureId {
-            id: self.identity.process(id_in, self.backend),
-            data: &self.data,
-        }
-    }
-
-    /// Acquire read access to this `Registry`'s contents.
-    ///
-    /// The caller must present a mutable reference to a `Token<A>`,
-    /// for some type `A` that comes before this `Registry`'s resource
-    /// type `T` in the lock ordering. A `Token<Root>` grants
-    /// permission to lock any field; see [`Token::root`].
-    ///
-    /// Once the read lock is acquired, return a new `Token<T>`, along
-    /// with a read guard for this `Registry`'s [`Storage`], which can
-    /// be indexed by id to get at the actual resources.
-    ///
-    /// The borrow checker ensures that the caller cannot again access
-    /// its `Token<A>` until it has dropped both the guard and the
-    /// `Token<T>`.
-    ///
-    /// See the [`Hub`] type for more details on locking.
-    pub(crate) fn read<'a, A: Access<T>>(
-        &'a self,
-        _token: &'a mut Token<A>,
-    ) -> (RwLockReadGuard<'a, Storage<T, I>>, Token<'a, T>) {
-        (self.data.read(), Token::new())
-    }
-
-    /// Acquire write access to this `Registry`'s contents.
-    ///
-    /// The caller must present a mutable reference to a `Token<A>`,
-    /// for some type `A` that comes before this `Registry`'s resource
-    /// type `T` in the lock ordering. A `Token<Root>` grants
-    /// permission to lock any field; see [`Token::root`].
-    ///
-    /// Once the lock is acquired, return a new `Token<T>`, along with
-    /// a write guard for this `Registry`'s [`Storage`], which can be
-    /// indexed by id to get at the actual resources.
-    ///
-    /// The borrow checker ensures that the caller cannot again access
-    /// its `Token<A>` until it has dropped both the guard and the
-    /// `Token<T>`.
-    ///
-    /// See the [`Hub`] type for more details on locking.
-    pub(crate) fn write<'a, A: Access<T>>(
-        &'a self,
-        _token: &'a mut Token<A>,
-    ) -> (RwLockWriteGuard<'a, Storage<T, I>>, Token<'a, T>) {
-        (self.data.write(), Token::new())
-    }
-
-    /// Unregister the resource at `id`.
-    ///
-    /// The caller must prove that it already holds a write lock for
-    /// this `Registry` by passing a mutable reference to this
-    /// `Registry`'s storage, obtained from the write guard returned
-    /// by a previous call to [`write`], as the `guard` parameter.
-    pub fn unregister_locked(&self, id: I, guard: &mut Storage<T, I>) -> Option<T> {
-        let value = guard.remove(id);
-        //Note: careful about the order here!
-        self.identity.free(id);
-        //Returning None is legal if it's an error ID
-        value
-    }
-
-    /// Unregister the resource at `id` and return its value, if any.
-    ///
-    /// The caller must present a mutable reference to a `Token<A>`,
-    /// for some type `A` that comes before this `Registry`'s resource
-    /// type `T` in the lock ordering.
-    ///
-    /// This returns a `Token<T>`, but it's almost useless, because it
-    /// doesn't return a lock guard to go with it: its only effect is
-    /// to make the token you passed to this function inaccessible.
-    /// However, the `Token<T>` can be used to satisfy some functions'
-    /// bureacratic expectations that you will have one available.
-    ///
-    /// The borrow checker ensures that the caller cannot again access
-    /// its `Token<A>` until it has dropped both the guard and the
-    /// `Token<T>`.
-    ///
-    /// See the [`Hub`] type for more details on locking.
-    pub(crate) fn unregister<'a, A: Access<T>>(
-        &self,
-        id: I,
-        _token: &'a mut Token<A>,
-    ) -> (Option<T>, Token<'a, T>) {
-        let value = self.data.write().remove(id);
-        //Note: careful about the order here!
-        self.identity.free(id);
-        //Returning None is legal if it's an error ID
-        (value, Token::new())
-    }
-
-    pub fn label_for_resource(&self, id: I) -> String {
-        let guard = self.data.read();
-
-        let type_name = guard.kind;
-        match guard.get(id) {
-            Ok(res) => {
-                let label = res.label();
-                if label.is_empty() {
-                    format!("<{}-{:?}>", type_name, id.unzip())
-                } else {
-                    label.to_string()
-                }
-            }
-            Err(_) => format!(
-                "<Invalid-{} label={}>",
-                type_name,
-                guard.label_for_invalid_id(id)
-            ),
-        }
     }
 }
